+++ src/main/java/de/theidler/create_mobile_packages/blocks/bee_port/BeePortBlockEntity.java
@@ -363,7 +363,7 @@ public class BeePortBlockEntity extends PackagePortBlockEntity {
         // Check if the item can be sent to another drone port.
         if (CMPConfigs.server().portToPort.get() && !PackageItem.matchAddress(address, addressFilter)) {
             BeePortBlockEntity beePortBlockEntity = CMPHelper.getClosestBeePort(level, address, this.getBlockPos(), null, getLogisticsNetworkId());
-            if (beePortBlockEntity != null && !beePortBlockEntity.isFull()) {
+            if (beePortBlockEntity != null && beePortBlockEntity.hasSpaceForPackageAndRobo()) {
                 sendDrone(itemStack, slot);
             }
         }
@@ -524,36 +524,35 @@ public class BeePortBlockEntity extends PackagePortBlockEntity {
     }
 
     /**
-     * Checks if the drone port is full, considering a specified number of slots to leave empty.
+     * Checks if the at least one of the PackageSlots is empty.
      *
-     * @param slotsToLeaveEmpty The number of slots that should remain empty.
-     * @return True if the number of empty slots is less than or equal to the specified slots to leave empty, false otherwise.
+     * @return True if at least one slot is empty, false if all slots are full.
      */
-    public boolean hasFullInventory(int slotsToLeaveEmpty) {
-        int emptySlots = 0;
+    public boolean hasSpaceForPackage() {
         for (int i = 0; i < inventory.getSlots(); i++) {
             if (inventory.getStackInSlot(i).isEmpty()) {
-                emptySlots++;
+                return true;
             }
         }
-        return emptySlots <= slotsToLeaveEmpty;
-    }
-
-    public synchronized boolean hasFullRoboSlot(int leaveEmpty) {
-        return roboBeeInventory.getStackInSlot(0).getCount() >= ROBOBEE_INVENTORY_STACK_SIZE - leaveEmpty;
+        return false;
     }
 
     /**
-     * Checks if the drone port is full.
+     * Checks if there is space for at least one Robo-Bee in the RoboBeeInventory.
      *
-     * @return True if the drone port is full, false otherwise.
+     * @return True if there is space for at least one Robo-Bee, false otherwise.
      */
-    public boolean isFull() {
-        return isFull(0);
+    public synchronized boolean hasSpaceForRobo() {
+        return roboBeeInventory.getStackInSlot(0).getCount() < ROBOBEE_INVENTORY_STACK_SIZE;
     }
 
-    public boolean isFull(int slotsToLeaveEmpty) {
-        return hasFullInventory(slotsToLeaveEmpty) || hasFullRoboSlot(0);
+    /**
+     * Checks if there is space for at least one Package and one Robo-Bee in the respective inventories.
+     *
+     * @return True if there is space for at least one Package and one Robo-Bee, false otherwise.
+     */
+    public boolean hasSpaceForPackageAndRobo() {
+        return hasSpaceForPackage() && hasSpaceForRobo();
     }
 
     /**
@@ -565,9 +564,9 @@ public class BeePortBlockEntity extends PackagePortBlockEntity {
      */
     public synchronized boolean canAcceptEntity(VirtualRobo entity, Boolean hasPackage) {
         if (this.isRemoved()) return false;
-        if (entity == null) return hasPackage ? !isFull() : !hasFullRoboSlot(0);
+        if (entity == null) return hasPackage ? hasSpaceForPackageAndRobo() : hasSpaceForRobo();
         if (hasRoboRequest()) return false;
-        return hasPackage ? !isFull() : !hasFullRoboSlot(0);
+        return hasPackage ? hasSpaceForPackageAndRobo() : hasSpaceForRobo();
     }
 
     public ItemStackHandler getRoboBeeInventory() {