+++ src/main/java/de/theidler/create_mobile_packages/robo/RoboTarget.java
@@ -20,7 +20,7 @@ public interface RoboTarget {
         return null;
     }
 
-    default boolean isValid() {
+    default boolean isValid(VirtualRobo robo) {
         return true;
     }